# Transform.insertNodes & Transform.insertFragment performance optimize