--- a/packages/slate/src/core/apply.ts
+++ b/packages/slate/src/core/apply.ts
@@ -1,11 +1,13 @@
 import { PathRef } from '../interfaces/path-ref'
 import { PointRef } from '../interfaces/point-ref'
 import { RangeRef } from '../interfaces/range-ref'
-import { DIRTY_PATH_KEYS, DIRTY_PATHS, FLUSHING } from '../utils/weak-maps'
+import { FLUSHING } from '../utils/weak-maps'
 import { Path } from '../interfaces/path'
 import { Transforms } from '../interfaces/transforms'
 import { WithEditorFirstArg } from '../utils/types'
 import { Editor } from '../interfaces/editor'
+import { isBatchingDirtyPaths } from './batch-dirty-paths'
+import { updateDirtyPaths } from './update-dirty-paths'
 
 export const apply: WithEditorFirstArg<Editor['apply']> = (editor, op) => {
   for (const ref of Editor.pathRefs(editor)) {
@@ -20,41 +22,14 @@ export const apply: WithEditorFirstArg<Editor['apply']> = (editor, op) => {
     RangeRef.transform(ref, op)
   }
 
-  const oldDirtyPaths = DIRTY_PATHS.get(editor) || []
-  const oldDirtyPathKeys = DIRTY_PATH_KEYS.get(editor) || new Set()
-  let dirtyPaths: Path[]
-  let dirtyPathKeys: Set<string>
-
-  const add = (path: Path | null) => {
-    if (path) {
-      const key = path.join(',')
-
-      if (!dirtyPathKeys.has(key)) {
-        dirtyPathKeys.add(key)
-        dirtyPaths.push(path)
-      }
-    }
+  // update dirty paths
+  if (!isBatchingDirtyPaths(editor)) {
+    const transform = Path.operationCanTransformPath(op)
+      ? (p: Path) => Path.transform(p, op)
+      : undefined
+    updateDirtyPaths(editor, editor.getDirtyPaths(op), transform)
   }
 
-  if (Path.operationCanTransformPath(op)) {
-    dirtyPaths = []
-    dirtyPathKeys = new Set()
-    for (const path of oldDirtyPaths) {
-      const newPath = Path.transform(path, op)
-      add(newPath)
-    }
-  } else {
-    dirtyPaths = oldDirtyPaths
-    dirtyPathKeys = oldDirtyPathKeys
-  }
-
-  const newDirtyPaths = editor.getDirtyPaths(op)
-  for (const path of newDirtyPaths) {
-    add(path)
-  }
-
-  DIRTY_PATHS.set(editor, dirtyPaths)
-  DIRTY_PATH_KEYS.set(editor, dirtyPathKeys)
   Transforms.transform(editor, op)
   editor.operations.push(op)
   Editor.normalize(editor, {