--- a/packages/slate/src/interfaces/transforms/node.ts
+++ b/packages/slate/src/interfaces/transforms/node.ts
@@ -9,6 +9,7 @@ export interface NodeInsertNodesOptions<T extends Node> {
   hanging?: boolean
   select?: boolean
   voids?: boolean
+  batchDirty?: boolean
 }
 
 export interface NodeTransforms {