--- a/packages/slate/src/interfaces/transforms/text.ts
+++ b/packages/slate/src/interfaces/transforms/text.ts
@@ -15,6 +15,7 @@ export interface TextInsertFragmentOptions {
   at?: Location
   hanging?: boolean
   voids?: boolean
+  batchDirty?: boolean
 }
 
 export interface TextInsertTextOptions {