--- a/packages/slate/src/transforms-node/insert-nodes.ts
+++ b/packages/slate/src/transforms-node/insert-nodes.ts
@@ -8,6 +8,9 @@ import { Text } from '../interfaces/text'
 import { Element } from '../interfaces/element'
 import { Path } from '../interfaces/path'
 import { getDefaultInsertLocation } from '../utils'
+import { batchDirtyPaths } from '../core/batch-dirty-paths'
+import { BaseInsertNodeOperation } from '../interfaces'
+import { updateDirtyPaths } from '../core/update-dirty-paths'
 
 export const insertNodes: NodeTransforms['insertNodes'] = (
   editor,
@@ -89,12 +92,51 @@ export const insertNodes: NodeTransforms['insertNodes'] = (
       return
     }
 
-    for (const node of nodes) {
-      const path = parentPath.concat(index)
-      index++
-      editor.apply({ type: 'insert_node', path, node })
-      at = Path.next(at)
-    }
+    // PERF: batch update dirty paths
+    // batched ops used to transform existing dirty paths
+    const batchedOps: BaseInsertNodeOperation[] = []
+    const newDirtyPaths: Path[] = Path.levels(parentPath)
+    batchDirtyPaths(
+      editor,
+      () => {
+        for (const node of nodes as Node[]) {
+          const path = parentPath.concat(index)
+          index++
+
+          const op: BaseInsertNodeOperation = {
+            type: 'insert_node',
+            path,
+            node,
+          }
+          editor.apply(op)
+          at = Path.next(at as Path)
+
+          batchedOps.push(op)
+          if (!Text.isText) {
+            newDirtyPaths.push(path)
+          } else {
+            newDirtyPaths.push(
+              ...Array.from(Node.nodes(node), ([, p]) => path.concat(p))
+            )
+          }
+        }
+      },
+      () => {
+        updateDirtyPaths(editor, newDirtyPaths, p => {
+          let newPath: Path | null = p
+          for (const op of batchedOps) {
+            if (Path.operationCanTransformPath(op)) {
+              newPath = Path.transform(newPath, op)
+              if (!newPath) {
+                return null
+              }
+            }
+          }
+          return newPath
+        })
+      }
+    )
+
     at = Path.previous(at)
 
     if (select) {