--- a/packages/slate/test/transforms/insertFragment/of-blocks/block-empty.tsx
+++ b/packages/slate/test/transforms/insertFragment/of-blocks/block-empty.tsx
@@ -2,7 +2,7 @@
 import { Transforms } from 'slate'
 import { jsx } from '../../..'
 
-export const run = editor => {
+export const run = (editor, options = {}) => {
   Transforms.insertFragment(
     editor,
     <fragment>
@@ -11,7 +11,8 @@ export const run = editor => {
       </block>
       <block>two</block>
       <block>three</block>
-    </fragment>
+    </fragment>,
+    options
   )
 }
 export const input = (