--- a/packages/slate/test/transforms/insertFragment/of-blocks/block-nested.tsx
+++ b/packages/slate/test/transforms/insertFragment/of-blocks/block-nested.tsx
@@ -2,14 +2,15 @@
 import { Transforms } from 'slate'
 import { jsx } from '../../..'
 
-export const run = editor => {
+export const run = (editor, options = {}) => {
   Transforms.insertFragment(
     editor,
     <fragment>
       <block>one</block>
       <block>two</block>
       <block>three</block>
-    </fragment>
+    </fragment>,
+    options
   )
 }
 export const input = (