--- a/packages/slate/test/transforms/insertFragment/of-blocks/with-inline.tsx
+++ b/packages/slate/test/transforms/insertFragment/of-blocks/with-inline.tsx
@@ -2,7 +2,7 @@
 import { Transforms } from 'slate'
 import { jsx } from '../../..'
 
-export const run = editor => {
+export const run = (editor, options = {}) => {
   Transforms.insertFragment(
     editor,
     <fragment>
@@ -15,7 +15,8 @@ export const run = editor => {
       <block>
         seven<inline>eight</inline>nine
       </block>
-    </fragment>
+    </fragment>,
+    options
   )
 }
 export const input = (