--- a/packages/slate/test/transforms/insertFragment/of-mixed/block-empty2.tsx
+++ b/packages/slate/test/transforms/insertFragment/of-mixed/block-empty2.tsx
@@ -2,13 +2,14 @@
 import { Transforms } from 'slate'
 import { jsx } from '../../..'
 
-export const run = editor => {
+export const run = (editor, options = {}) => {
   Transforms.insertFragment(
     editor,
     <fragment>
       <block>two</block>
       <text>three</text>
-    </fragment>
+    </fragment>,
+    options
   )
 }
 export const input = (