--- a/packages/slate/test/transforms/insertFragment/of-mixed/block-empty3.tsx
+++ b/packages/slate/test/transforms/insertFragment/of-mixed/block-empty3.tsx
@@ -2,13 +2,14 @@
 import { Transforms } from 'slate'
 import { jsx } from '../../..'
 
-export const run = editor => {
+export const run = (editor, options = {}) => {
   Transforms.insertFragment(
     editor,
     <fragment>
       <text>one</text>
       <block>two</block>
-    </fragment>
+    </fragment>,
+    options
   )
 }
 export const input = (