--- a/packages/slate/test/transforms/insertFragment/of-mixed/block-start.tsx
+++ b/packages/slate/test/transforms/insertFragment/of-mixed/block-start.tsx
@@ -2,14 +2,15 @@
 import { Transforms } from 'slate'
 import { jsx } from '../../..'
 
-export const run = editor => {
+export const run = (editor, options = {}) => {
   Transforms.insertFragment(
     editor,
     <fragment>
       <text>one</text>
       <block>two</block>
       <text>three</text>
-    </fragment>
+    </fragment>,
+    options
   )
 }
 export const input = (