--- a/packages/slate/test/transforms/insertFragment/of-tables/merge-cells-with-nested-blocks.tsx
+++ b/packages/slate/test/transforms/insertFragment/of-tables/merge-cells-with-nested-blocks.tsx
@@ -2,7 +2,7 @@
 import { Transforms } from 'slate'
 import { jsx } from '../../..'
 
-export const run = editor => {
+export const run = (editor, options = {}) => {
   Transforms.insertFragment(
     editor,
     <block>
@@ -16,7 +16,8 @@ export const run = editor => {
           </block>
         </block>
       </block>
-    </block>
+    </block>,
+    options
   )
 }
 export const input = (