--- a/packages/slate/test/transforms/insertFragment/of-tables/merge-into-full-cells.tsx
+++ b/packages/slate/test/transforms/insertFragment/of-tables/merge-into-full-cells.tsx
@@ -2,7 +2,7 @@
 import { Transforms } from 'slate'
 import { jsx } from '../../..'
 
-export const run = editor => {
+export const run = (editor, options = {}) => {
   Transforms.insertFragment(
     editor,
     <block>
@@ -12,7 +12,8 @@ export const run = editor => {
           <block>New 2</block>
         </block>
       </block>
-    </block>
+    </block>,
+    options
   )
 }
 export const input = (