--- a/packages/slate/test/transforms/insertFragment/of-texts/inline-middle.tsx
+++ b/packages/slate/test/transforms/insertFragment/of-texts/inline-middle.tsx
@@ -2,8 +2,8 @@
 import { Transforms } from 'slate'
 import { jsx } from '../../..'
 
-export const run = editor => {
-  Transforms.insertFragment(editor, <fragment>fragment</fragment>)
+export const run = (editor, options = {}) => {
+  Transforms.insertFragment(editor, <fragment>fragment</fragment>, options)
 }
 export const input = (
   <editor>