--- a/packages/slate/test/transforms/insertFragment/of-texts/with-multiple.tsx
+++ b/packages/slate/test/transforms/insertFragment/of-texts/with-multiple.tsx
@@ -2,13 +2,14 @@
 import { Transforms } from 'slate'
 import { jsx } from '../../..'
 
-export const run = editor => {
+export const run = (editor, options = {}) => {
   Transforms.insertFragment(
     editor,
     <fragment>
       <text>one</text>
       <text>two</text>
-    </fragment>
+    </fragment>,
+    options
   )
 }
 export const input = (