--- a/packages/slate/test/transforms/insertNodes/block/block-empty.tsx
+++ b/packages/slate/test/transforms/insertNodes/block/block-empty.tsx
@@ -10,12 +10,13 @@ export const input = (
     <block>not empty</block>
   </editor>
 )
-export const run = editor => {
+export const run = (editor, options = {}) => {
   Transforms.insertNodes(
     editor,
     <block>
       <text />
-    </block>
+    </block>,
+    options
   )
 }
 export const output = (