--- a/packages/slate/test/transforms/insertNodes/block/block-middle.tsx
+++ b/packages/slate/test/transforms/insertNodes/block/block-middle.tsx
@@ -2,12 +2,13 @@
 import { Transforms } from 'slate'
 import { jsx } from '../../..'
 
-export const run = editor => {
+export const run = (editor, options = {}) => {
   Transforms.insertNodes(
     editor,
     <block>
       <text />
-    </block>
+    </block>,
+    options
   )
 }
 export const input = (