… .changeset/witty-apples-return.md
… packages/slate/test/transforms/insertNodes/block/inline-void.tsx
… packages/slate/test/transforms/insertNodes/inline/block-empty.tsx
… packages/slate/test/transforms/insertNodes/inline/block-end.tsx
… packages/slate/test/transforms/insertNodes/inline/block-middle.tsx
… packages/slate/test/transforms/insertNodes/inline/block-start.tsx
… packages/slate/test/transforms/insertNodes/inline/block-void.tsx
… packages/slate/test/transforms/insertNodes/inline/inline-middle.tsx
… packages/slate/test/transforms/insertNodes/path/block.tsx
… packages/slate/test/transforms/insertNodes/path/inline.tsx
… packages/slate/test/transforms/insertNodes/path/multiple-inline-not-end.tsx
… packages/slate/test/transforms/insertNodes/path/multiple-inline.tsx
… packages/slate/test/transforms/insertNodes/path/multiple.tsx
… packages/slate/test/transforms/insertNodes/path/text.tsx
… packages/slate/test/transforms/insertNodes/select-true/block.tsx
… packages/slate/test/transforms/insertNodes/selection/none-empty.tsx
… packages/slate/test/transforms/insertNodes/selection/none-end.tsx
… packages/slate/test/transforms/insertNodes/void/at-path.tsx
… packages/slate/test/transforms/insertNodes/void/block-nested.tsx
… packages/slate/test/transforms/insertNodes/void/block.tsx
… packages/slate/test/transforms/insertNodes/void/inline.tsx
… packages/slate/test/transforms/insertNodes/voids-true/block.tsx
… packages/slate/test/transforms/insertNodes/voids-true/inline.tsx