@mainhanu 2023-11-07
added `batchDirty` to insertNodes & insertFragment options for unit test to make sure: with or without batchDirty, the normalization paths is the same